% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_morlet.R
\name{analytic.ridge}
\alias{analytic.ridge}
\title{Analytic Ridge Extraction}
\usage{
analytic.ridge(wave, ridge_only = FALSE, plim = c(-Inf, +Inf))
}
\arguments{
\item{wave}{Wavelet object.}

\item{ridge_only}{Logical. Return only ridge if TRUE.}

\item{plim}{Numeric vector. Period limits.}
}
\value{
Ridge or modified wavelet object.
}
\description{
Performs analytic ridge extraction from a wavelet transform.
}
