% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_ts.R
\name{approx_ts}
\alias{approx_ts}
\title{Produce an evenly-spaced time series by simple interpolation}
\usage{
approx_ts(
  data,
  tcoord = "x",
  dcoord = "y",
  scale = 1,
  n = 2048,
  thin = FALSE,
  spline = FALSE,
  xlim = NULL
)
}
\arguments{
\item{data}{The data input, typically a dataframe with two columns}

\item{tcoord}{The name of the column name of the time coordinate. Defaults to "x"}

\item{dcoord}{The name of the column name of the data coordinate. Defaults to "y"}

\item{scale}{Time scaling (the coordinate "tcoord" will be divided by `scale`).}

\item{n}{(defaults to 2048) is the number of data points in the output}

\item{thin}{Discards data that are sampled with a time space less than 0.5 times the time resolution in the output}

\item{spline}{If TRUE, the standard `spline` routine will be used, otherwise, `approx`}

\item{xlim}{If provided, brackets the time output}
}
\value{
a time series R object
}
\description{
Uses standard R routines 'approx' or 'spline' to generate an evenly-spaced time series
based on data supplied on a time coordinate grid.
}
\author{
Michel Crucifix
}
