% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arspec.R
\name{arspec}
\alias{arspec}
\title{Theoretical Spectrum of autoregressive spectrum}
\usage{
arspec(f, ar)
}
\description{
Calculates the theoretical density frequency of an
autoregressive process, according to equation
(3.3) of Akaike (1969)
}
