% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{cis}
\alias{cis}
\title{Trigonometric exponential function}
\usage{
cis(x)
}
\arguments{
\item{x}{Numeric input value.}
}
\value{
Complex exponential of x.
}
\description{
Returns the complex exponential of x.
}
