% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_morlet.R
\docType{data}
\name{col_wavelet}
\alias{col_wavelet}
\title{Continuous Morlet Wavelet Transform and Related Methods}
\format{
An object of class \code{character} of length 100.
}
\usage{
col_wavelet
}
\description{
This file provides functions for performing the Continuous Morlet Wavelet Transform (CWT)
on input time series, visualizing the results, extracting ridges, and reconstructing time series.
}
\details{
References:
Mallat, S. 1998: A wavelet Tour of Signal Processing. Academic Press, 577 pp.
Tchawitchia P., Wavelets, functions and operators, ch. 3 in Wavelets : theory and applications,
Erlenbacher et al. eds, Oxford University Press 1996 [for ridge extraction]
}
\keyword{internal}
