% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_morlet.R
\name{cross_morlet}
\alias{cross_morlet}
\title{Continuous Cross Morlet Wavelet Transform}
\usage{
cross_morlet(A, B, ...)
}
\arguments{
\item{A}{First time series.}

\item{B}{Second time series.}

\item{...}{Additional parameters for Morlet transform.}
}
\value{
Cross wavelet transform matrix.
}
\description{
Computes the cross wavelet transform between two time series.
}
