% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_morlet.R
\name{cwt_morlet}
\alias{cwt_morlet}
\title{Continuous Morlet Wavelet Transform}
\usage{
cwt_morlet(
  A,
  inter = 20,
  k0 = 5.6,
  amin = 1,
  amax = Inf,
  calcmask = TRUE,
  scale = NA,
  deriv = FALSE
)
}
\arguments{
\item{A}{Numeric or time series vector.}

\item{inter}{Integer. Number of intervals per octave.}

\item{k0}{Numeric. Morlet wavelet parameter.}

\item{amin}{Minimum scale.}

\item{amax}{Maximum scale.}

\item{calcmask}{Logical. Whether to mask edge effects.}

\item{scale}{Numeric. Specific scale(s) to use.}

\item{deriv}{Logical. Whether to calculate derivatives.}
}
\value{
A wavelet object (matrix with attributes).
}
\description{
Computes the Continuous Morlet Wavelet Transform (CWT) of a time series.
}
\details{
Returns a wavelet object containing the complex CWT coefficients, with attributes for time, scale, period, and more.
}
