% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{develop.discreteSpectrum}
\alias{develop.discreteSpectrum}
\title{Discrete spectrum reconstruction}
\usage{
\method{develop}{discreteSpectrum}(M, start = NULL, end = NULL, deltat = NULL, times = NULL, ...)
}
\arguments{
\item{M}{discreteSpectrum object}

\item{start}{see \code{develop}}

\item{end}{see \code{develop}}

\item{deltat}{see \code{develop}}

\item{times}{see \code{develop}}

\item{...}{Further arguments passed through. May include:
- `dfunction`: trigonometrical function (`cos`, `sin`, or `cis`)
- `max_freq`: maximum number of frequencies to include
- `sum`: logical; sum components in reconstruction?
- `trendshift`: logical; account for trend and shift encoded in the object?}
}
\value{
List of reconstructed components if sum=FALSE, full reconstructed time series otherwise.
}
\description{
Reconstruct a time series from a discreteSpectrum object.
}
\note{
If none of times, start and deltat are supplied, will reconstruct based on the attribute `xdata`.
}
