% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{develop}
\alias{develop}
\title{Develop a spectrum into a time series (generic)}
\usage{
develop(M, start = NULL, end = NULL, deltat = NULL, times = NULL, ...)
}
\arguments{
\item{M}{input object (typically a spectrum or decomposition)}

\item{start}{If supplied, the start time of the decomposition (overrides `times`)}

\item{end}{If supplied, the end time of the decomposition (used with `start` and `deltat`)}

\item{deltat}{If supplied, the time interval (used with `start` and `end`)}

\item{times}{If supplied, explicit times for the decomposition}

\item{...}{Further arguments passed to methods}
}
\value{
Nothing; method dispatches to appropriate method.
}
\description{
This is a generic function to reconstruct a time series from a spectrum decomposition.
}
\note{
Place holder for type-specific develop functions.
}
