% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfft_support.R
\name{mfft_anova}
\alias{mfft_anova}
\alias{as.data.frame.discreteSpectrum}
\alias{plot.discreteSpectrum}
\alias{lines.discreteSpectrum}
\alias{print.discreteSpectrum}
\title{MFFT ANOVA
not ready. do not use.}
\usage{
mfft_anova(M)

\method{as.data.frame}{discreteSpectrum}(x)

\method{plot}{discreteSpectrum}(M, periods = FALSE, labels = NULL, ...)

\method{lines}{discreteSpectrum}(M, ...)

\method{print}{discreteSpectrum}(M, ...)
}
\arguments{
\item{periods}{if TRUE will add a lower axis with period labels}

\item{labels}{to be set above the frequency peaks. Can be the output of `attributeTone`}

\item{a}{`discreteSpectrum` object, typically the output of a `mfft` call.}
}
\description{
MFFT ANOVA
not ready. do not use.
}
