% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_morlet.R
\name{enhance_ridge}
\alias{enhance_ridge}
\title{Enhance Ridge}
\usage{
enhance_ridge(A, R, confidence = 0.95, inter = 0.08)
}
\arguments{
\item{A}{Input time series.}

\item{R}{Ridge object.}

\item{confidence}{Confidence interval for enhancement.}

\item{inter}{Step size for enhancement.}
}
\value{
Enhanced ridge amplitude vector.
}
\description{
Enhances a previously extracted ridge.
}
