% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haar.R
\name{haar}
\alias{haar}
\title{Haar fluctuation spectrum}
\usage{
haar(x, q = 2, discarded_scales = 4)
}
\arguments{
\item{x}{the input series (typically a numeric)}

\item{q}{the fluctuation order}
}
\description{
Haar fluctuation spectrum after Lovejoy
still needs to be documented
currently requires n=2^i data points
}
\examples{
x = rnorm(2048)
xi1 = haar(x, q=1)
xi2 = haar(x, q=2)
}
