% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R, R/data_description.R
\docType{data}
\name{harmonic_sample}
\alias{harmonic_sample}
\title{harmonic_sample}
\format{
A list of class \code{discreteSpectrum} with the following components:
\describe{
  \item{Amp}{Vector of amplitudes of the harmonic components}
  \item{Freq}{Vector of frequencies of the harmonic components}
  \item{Phases}{Vector of phases of the harmonic components}
}
The raw time series data are stored as an attribute "data" (a \code{ts} object).
}
\usage{
harmonic_sample

data(harmonic_sample)
}
\description{
A simple harmonic time series with multiple frequency components.

A simple time series for testing purposes
cos(t*0.13423167+0.00) + 1.3 * cos(t*0.119432+2.314) + 0.134994 + 0.4*cos(t*0.653167) + 0.11 * cos(t*0.78913498) 
for t from 0 to 1023. It is coded as a `discreteSpectrum` objcet and self-describes its spectral decomposition.
}
\keyword{datasets}
