% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R, R/data_description.R
\docType{data}
\name{harmonic_sample_noisy}
\alias{harmonic_sample_noisy}
\title{harmonic_sample_noisy}
\format{
Same as \code{harmonic_sample}, with added Gaussian noise in the "data" attribute.
}
\usage{
harmonic_sample_noisy

data(harmonic_sample_noisy)
}
\description{
A noisy version of `harmonic_sample`.

Same has harmonic sample but with an added random time series
cos(t*0.13423167+0.00) + 1.3 * cos(t*0.119432+2.314) + 0.134994 + 0.4*cos(t*0.653167) + 0.11 * cos(t*0.78913498) + rnorm(1024)*0.12
for t from 0 to 1023. It is coded as a `discreteSpectrum` objcet and self-describes its spectral decomposition.
}
\keyword{datasets}
