% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilbert.R
\name{hilbert_extension}
\alias{hilbert_extension}
\title{Hilbert extension}
\usage{
hilbert_extension(x)
}
\arguments{
\item{x}{Input array}
}
\value{
Analytic signal, of length \code{n}, returned as a complex vector or
}
\description{
Computes the extension of a real valued signal to a complex signas with
the Hilbert transform
}
