% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{insol_scaled}
\alias{insol_scaled}
\title{insol_scaled}
\format{
A \code{ts} object of length 4096.
}
\usage{
insol_scaled
}
\description{
Scaled test insolation dataset used for time series analysis.
}
\keyword{datasets}
