% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{insol_truncated}
\alias{insol_truncated}
\title{insol_truncated}
\format{
A \code{ts} object of length 4096.
}
\usage{
insol_truncated
}
\description{
Truncated/thresholded version of \code{insol_scaled}, values in \{-1, 0, 1\}.
}
\keyword{datasets}
