% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{lines.discreteSpectrum}
\alias{lines.discreteSpectrum}
\title{Add lines to discrete spectrum plot}
\usage{
\method{lines}{discreteSpectrum}(M, ...)
}
\arguments{
\item{M}{discreteSpectrum object}

\item{...}{Further graphical arguments}
}
\description{
Adds lines and points to the spectrum plot for a discreteSpectrum object.
}
