% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{mem}
\alias{mem}
\alias{plot.memObject}
\title{Maximum entropy method}
\usage{
mem(A, method = "burg", order.max = 90, ...)

\method{plot}{memObject}(x, period = FALSE, xaxp = NULL, yaxp = NULL, ...)
}
\arguments{
\item{A}{input, a time-series object}

\item{method}{Estimation method, passed to the standard `ar` function, which is one of "yule-walker", "burg", "ols", "mle", "yw". Defaults to "burg".}

\item{order.max}{maximum order of the autoregressive model being fitted}

\item{...}{Additional arguments passed to \code{ar}.}

\item{x}{A \code{memObject}, the result from \code{mem()} (for plotting).}

\item{period}{Logical. If \code{TRUE}, the x axis is shown as period instead of frequency.}

\item{xaxp}{Arguments passed to adjust the x axis.}

\item{yaxp}{Arguments passed to adjust the y axis.}
}
\value{
a memObject
}
\description{
Maximum entropy spectral analysis (MESA) is the statistical estimation of the power spectrum of a stationary time series using the maximum entropy (ME) method. The resulting ME spectral estimator is computed by fitting an autoregressive model to the input time series.
}
\references{
\insertRef{Burg75aa}{gtseries}
\insertRef{percival98}{gtseries}
\insertRef{Ghil02aa}{gtseries}
\insertRef{pardoiguzquiza06aa}{gtseries}
\insertRef{pardo-iguzquiza21aa}{gtseries}
}
