% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{plot.periodogram_complex}
\alias{plot.periodogram_complex}
\title{Plot a complex periodogram object}
\usage{
\method{plot}{periodogram_complex}(x, ...)
}
\arguments{
\item{x}{periodogram_complex object}

\item{...}{further arguments passed to the default plot method, may include:
\describe{
  \item{log}{log plot, e.g. \code{"x"}, \code{"y"}, or \code{"xy"} (default: \code{"y"})}
  \item{xlab}{x-axis label (default: \code{"Frequency"})}
  \item{ylab}{y-axis label (default: \code{"Power density"})}
  \item{type}{plot type, e.g. \code{"l"} for lines}
  \item{main}{plot title}
}}
}
\value{
The plot
}
\description{
Plot a complex periodogram object
}
