% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_randomize.R
\name{phase_randomize}
\alias{phase_randomize}
\title{Phase Randomizer}
\usage{
phase_randomize(x)
}
\arguments{
\item{x}{the input time series}
}
\value{
x the output time series.
}
\description{
randomizes the phases of a real signal, for non-linear time series analysis
It works by computing the FFT of the signal, then randomize the phases of all
complex numbers, but preserving the complex conjugate relationship guarantee
that the inverse transform will be real
}
