% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{plot.discreteSpectrum}
\alias{plot.discreteSpectrum}
\title{Plot discrete spectrum}
\usage{
\method{plot}{discreteSpectrum}(M, periods = FALSE, labels = NULL, ...)
}
\arguments{
\item{M}{discreteSpectrum object}

\item{periods}{If TRUE, will add a lower axis with period labels}

\item{labels}{Labels to be set above the frequency peaks}

\item{...}{Further graphical arguments}
}
\description{
Plots the amplitude spectrum of a discreteSpectrum object.
}
