% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_morlet.R
\name{powerspectrum.wavelet}
\alias{powerspectrum.wavelet}
\title{Power Spectrum of a Wavelet Transform}
\usage{
powerspectrum.wavelet(wave, ...)
}
\arguments{
\item{wave}{A wavelet object.}
}
\value{
Data frame with frequency and power columns.
}
\description{
Computes the mean power spectrum from a wavelet transform.
}
