% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_morlet.R
\name{reconstruct_morlet}
\alias{reconstruct_morlet}
\title{Reconstruct Time Series from Morlet Wavelet Transform}
\usage{
reconstruct_morlet(W, scales = c(-Inf, Inf), periods = NULL)
}
\arguments{
\item{W}{A wavelet object.}

\item{scales}{Numeric vector. Restrict reconstruction to these scales.}

\item{periods}{Numeric vector. Restrict to these periods.}
}
\value{
Time series object.
}
\description{
Reconstructs the original time series from its Morlet wavelet transform, optionally
restricting to certain scales or periods.
}
