% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_morlet.R
\name{search_ridge}
\alias{search_ridge}
\title{Ridge Extraction from CWT}
\usage{
search_ridge(
  A,
  first_guess,
  k0 = 5.6,
  tol = 1e-06,
  itmax = 20,
  B = seq(along = A),
  window = c(0, Inf)
)
}
\arguments{
\item{A}{Numeric or time series vector.}

\item{first_guess}{Initial scale guess.}

\item{k0}{Morlet parameter.}

\item{tol}{Tolerance for convergence.}

\item{itmax}{Maximum iterations.}

\item{B}{Time index vector.}

\item{window}{Numeric vector for scale window.}
}
\value{
Ridge amplitude vector with attributes.
}
\description{
Extracts the ridge (maximal amplitude curve) from a continuous wavelet transform.
}
