% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toneCombinations.R
\name{toneCombinations}
\alias{toneCombinations}
\title{Generation of combination of tones}
\usage{
toneCombinations(omegas, keepPositives = TRUE)
}
\arguments{
\item{keepPositives}{: if TRUE, then only keeps positive combinations of frequencies}

\item{omegas:}{vector of references frequencies, optionally with rownames,}
}
\value{
a vector with combination of tones and explicit rownames, using, if available, the
        rownames provided in the input vector omega
}
\description{
Generates a vector with combinations of an input vector of frequencies, wih
explicit label names, up to order 3 (this could be made more flexible is the future)
}
\examples{
omegas <- c( 0.123, 0.14312, 0.33251, 0.554313)
print(toneCombinations(omegas)
}
\author{
Michel Crucifix
}
