% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{as.data.frame.discreteSpectrum}
\alias{as.data.frame.discreteSpectrum}
\title{Convert discreteSpectrum object to data.frame}
\usage{
\method{as.data.frame}{discreteSpectrum}(x, row.names, optional, ...)
}
\arguments{
\item{x}{discreteSpectrum object}

\item{row.names}{kept for compatibility with as.data.frame S3 method}

\item{optional}{kept for compatibility with as.data.frame S3 method}

\item{...}{passed to the data.frame function}
}
\value{
Data.frame with columns Freq, Amp, Phases
}
\description{
Converts a discreteSpectrum object into a data.frame.
}
