% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_morlet.R
\name{cwt_morlet}
\alias{cwt_morlet}
\alias{plot.wavelet}
\alias{powerspectrum.wavelet}
\title{Continous Morlet Wavelet Transform}
\usage{
cwt_morlet(
  A,
  inter = 20,
  k0 = 5.6,
  amin = 1,
  amax = Inf,
  calcmask = TRUE,
  scale = NA,
  deriv = FALSE
)

\method{plot}{wavelet}(
  wave,
  resx = 400,
  resy = 300,
  xlab = "Time",
  ylab = "Period",
  scaling_correction = 0,
  col = col_wavelet,
  legend = FALSE,
  Mode = Mod,
  plotMask = TRUE,
  ...
)

powerspectrum.wavelet(wave, ...)
}
\description{
Continous Morlet Wavelet Transform
}
