% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{develop.discreteSpectrum}
\alias{develop.discreteSpectrum}
\title{Discrete spectrum reconstruction}
\usage{
\method{develop}{discreteSpectrum}(M, start = NULL, end = NULL, deltat = NULL, times = NULL, ...)
}
\arguments{
\item{M}{discreteSpectrum object}

\item{start}{see \code{develop}}

\item{end}{see \code{develop}}

\item{deltat}{see \code{develop}}

\item{times}{see \code{develop}}

\item{...}{arguments passed through; may include:
\describe{
  \item{dfunction}{trigonometrical function ('cos', 'sin', or 'cis')}
  \item{max_freq}{maximum frequency to include}
  \item{sum}{logical: sum components in reconstruction?}
  \item{trendshift}{logical: account for trend and shift encoded in the object?}
}}
}
\value{
list of reconstructed components if sum=FALSE, full reconstructed time series otherwise
}
\description{
Discrete spectrum reconstruction
}
\note{
if none of times, start and deltat are supplied, will reconstruct based on the attribute `xdata`
}
