% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{harmonic_sample}
\alias{harmonic_sample}
\title{harmonic_sample}
\usage{
data(harmonic_sample)
}
\description{
A simple time series for testing purposes
cos(t*0.13423167+0.00) + 1.3 * cos(t*0.119432+2.314) + 0.134994 + 0.4*cos(t*0.653167) + 0.11 * cos(t*0.78913498) 
for t from 0 to 1023. It is coded as a `mfft_deco` objcet and self-describes its spectral decomposition.
}
\keyword{datasets}
