% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfft.R
\name{mfft}
\alias{mfft}
\title{Modified Fourier transform}
\usage{
mfft(xdata, minfreq = NULL, maxfreq = NULL, flag = 1, nfreq = 30)
}
\arguments{
\item{xdata}{The data provided either as a time series (advised), or as a vector. 
may be complex}

\item{minfreq, maxfreq}{If provided, bracket the frequencies to be probed. Note this are
more precisely angular velocities (2\pi / period), expressed in time-inverse units
with the time resolution encoded in `xdata` if the latter is a time series.}

\item{flag}{:
     Modified Fourier Transform                  if   flag = 1;
     Frequency Modified Fourier Transform        if   flag = 2;
     FMFT with additional non-linear correction  if   flag = 3
(while the first algorithm is app. 3 times faster than the third one, 
the third algorithm should be in general much more precise).  
The computed frequencies are in the range given by minfreq and maxfreq.}

\item{nfreq}{is the number of frequencies returned, must be smaller that the length of  xdata.}
}
\value{
STILL NEED TO BE DESCRIBED
}
\description{
Implementation of the the Frequency Modified Fourier Transform 
(Sidlichovsky and Nesvorny 1997, Cel. Mech. 65, 137). 
Given a quasi--periodic complex signal X + iY, the algorithm 
estimates the frequencies (f_j), amplitudes (A_j) and phases 
(psi_j) in its decomposition.
}
\author{
Michel Crucifix for the R code, and David Nesvorny for most of the supporting C code doing the
actual computations
\insertRef{sidlichovsky97aa}{gtseries}
}
