% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfft.R
\name{mfft_complex}
\alias{mfft_complex}
\title{Frequency Modified Fourier transform for real and  complex series}
\usage{
mfft_complex(
  x_data,
  n_freq = 10,
  min_freq = NULL,
  max_freq = NULL,
  correction = 1,
  fast = TRUE
)
}
\arguments{
\item{x_data}{The data provided either as a time series (advised), or as a vector.}

\item{n_freq}{The number of frequencies returned, must be smaller than the length of `x_data`.}

\item{min_freq, max_freq}{If provided, bracket the angular frequencies to be probed.
Note: these are angular velocities (\eqn{2\pi / \mathrm{period}}), expressed in time-inverse units,
with the time resolution encoded in `x_data` if the latter is a time series.
The default for `min_freq` is 0, and for `max_freq` is \eqn{\pi}.}

\item{correction}{0: no frequency correction (equivalent to Laskar);
1: frequency correction using linear approximation;
2: frequency correction using synthetic data;
3: second order-correction using synthetic data (all documented in the Sidlichovsky and Nesvorny reference)}

\item{fast}{(default = TRUE) Uses analytical formulations for the crossproducts involving sines and cosines.
Note: this is not really faster because the bottleneck is actually the golden section search,
but more elegant.}
}
\value{
a `discreteSpectrum` object, based on a data.frame with columns "Freq", "Amp", and "Phases".
}
\description{
Iteratively extracts dominant frequencies from a real-valued time series
using a modified Fourier transform with frequency correction. The function
applies a Hanning window to the data, and amplitudes are renormalised through
a Gram–Schmidt procedure based on cross-products of Fourier components.
The cross-products can be computed either via analytical approximation
(faster, following Sidlichovsky and Nesvorny) or explicitly by numerical
integration.
}
\details{
If frequencies \code{nu} are supplied, the routine uses them directly;
otherwise, it iteratively maximises the Fourier transform using golden
section search. The
function is typically called from [mfft_real()] once, and possibly a second
time to refine amplitudes and frequencies iteratively.
}
\examples{

data(harmonic_sample)
data_to_analyse <- develop(harmonic_sample)
spec <- mfft_real(data_to_analyse)
print(spec)

}
\references{
\insertRef{sidlichovsky97aa}{gtseries}
}
\author{
Michel Crucifix
}
