% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfft_support.R
\name{reconstruct_mfft}
\alias{reconstruct_mfft}
\alias{mfft_anova}
\alias{as.data.frame.mfft_deco}
\alias{plot.mfft_deco}
\alias{lines.mfft_deco}
\alias{print.mfft_deco}
\title{MFFT reconstruction}
\usage{
reconstruct_mfft(M, sum = TRUE)

mfft_anova(M)

\method{as.data.frame}{mfft_deco}(x)

\method{plot}{mfft_deco}(M, periods = FALSE, labels = NULL, ...)

\method{lines}{mfft_deco}(M, ...)

\method{print}{mfft_deco}(M, ...)
}
\arguments{
\item{M}{: mfft_deco object}

\item{sum}{: TRUE if user wants to sum components in the reconstruction}

\item{periods}{if TRUE will add a lower axis with period labels}

\item{labels}{to be set above the frequency peaks. Can be the output of `attributeTone`}

\item{a}{`mfft_deco` object, typically the output of a `mfft` call.}
}
\value{
list of reconstructed components if sum=FALSE,  full
        reconstructed time series otherwise
}
\description{
MFFT reconstruction

MFFT ANOVA
not ready. do not use.
}
