% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfft_support.R
\name{reconstruct_mfft}
\alias{reconstruct_mfft}
\alias{mfft_anova}
\alias{as.data.frame.mfft_deco}
\alias{plot.mfft_deco}
\alias{lines.mfft_deco}
\alias{print.mfft_deco}
\title{MFFT reconstruction}
\usage{
reconstruct_mfft(M)

mfft_anova(M)

\method{as.data.frame}{mfft_deco}(x)

\method{plot}{mfft_deco}(M, periods = FALSE, ...)

\method{lines}{mfft_deco}(M, ...)

\method{print}{mfft_deco}(M, ...)
}
\description{
MFFT reconstruction

MFFT ANOVA
not ready. do not use.
}
