% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfft.R
\name{mmfft}
\alias{mmfft}
\alias{plot.mmfft}
\title{Moving Modified Fourier Transform}
\usage{
mmfft(x, seglength = length(x)\%/\%16, ...)

\method{plot}{mmfft}(x, ...)
}
\arguments{
\item{x}{The data provided either as a time series (advised), or as a vector. 
may be complex}

\item{seglength}{Length of moving segment. Both x length and seglenthm are adviced to be a power of 2 (faster)}

\item{...}{passed to mmft}
}
\value{
a `mmfft` object
}
\description{
Moving Modified Fourier Transform
}
\note{
in the current implementation, the right-hand-side of x that is left after an integer division 
       by seglenth is discarded
}
\author{
Michel Crucifix
}
