% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\alias{plot.periodogram}
\title{Periodogram}
\usage{
periodogram(xdata)

\method{plot}{periodogram}(X, log = "xy", xlabel = "Frequency", ylabel = "Power density", ...)
}
\arguments{
\item{xdata}{}
}
\description{
Simple periodogram
}
\author{
Michel Crucifix for the R code
}
