% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\alias{plot.periodogram}
\title{Periodogram}
\usage{
periodogram(x)

\method{plot}{periodogram}(x, ...)
}
\arguments{
\item{x}{periodogram object}

\item{...}{further arguments passed to the default plot method, may include:
\describe{
  \item{log}{log plot.  axis to be logged, e.g. \code{"x"}, \code{"y"}, or \code{"xy"} (default: \code{"xy"})}
  \item{xlab}{x-axis label (default: \code{"Frequency"})}
  \item{ylab}{y-axis label (default: \code{"Power density"})}
  \item{type}{plot type, e.g. \code{"l"} for lines}
  \item{main}{plot title}
}}
}
\value{
The plot
}
\description{
Simple periodogram
}
\author{
Michel Crucifix for the R code
}
