% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{plot.discreteSpectrum}
\alias{plot.discreteSpectrum}
\title{Plot discrete spectrum}
\usage{
\method{plot}{discreteSpectrum}(x, ...)
}
\arguments{
\item{x}{discreteSpectrum object}

\item{...}{Further arguments passed through. May include:
- `periods`: Add a period axis bar below the frequency axis. Defaults to FALSE
- `labels`: if set, is passed to the standard plot function}
}
\description{
Plots the amplitude spectrum of a discreteSpectrum object.
}
