% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{print.discreteSpectrum}
\alias{print.discreteSpectrum}
\title{Print discrete spectrum}
\usage{
\method{print}{discreteSpectrum}(x, ...)
}
\arguments{
\item{x}{discreteSpectrum object}

\item{...}{Further arguments}
}
\description{
Prints a summary of the discreteSpectrum object.
}
