% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R, R/data_description.R
\docType{data}
\name{pseudo_log}
\alias{pseudo_log}
\title{pseudo_log}
\format{
A list of class \code{discreteSpectrum} with amplitudes, frequencies, and phases.
The raw binary time series is stored in the "data" attribute.
}
\usage{
pseudo_log

data(pseudo_log)
}
\description{
A pseudo-log dataset simulating a simple lithological record.

A truncated version (squared-shape) version of a harmonic signal for mimicking a sedimentary log
sign( cos(t*0.13423167+0.00) + 1.3 * cos(t*0.119432+2.314) + 0.134994 + 0.4*cos(t*0.653167) + 0.11 * cos(t*0.78913498) + 0.8))
for t from 0 to 8191. It is coded as a `discreteSpectrum` objcet and self-describes its spectral decomposition.
}
\keyword{datasets}
