% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa}
\alias{ssa}
\alias{plot.SSAObject}
\title{Singular spectrum analysis}
\usage{
ssa(X = X, N = length(X), M = M, Nrec = 10)

\method{plot}{SSAObject}(SSAObject, ...)
}
\arguments{
\item{X}{the input series (typically a numeric)}

\item{N}{if provided,  the length of the input series to beconsiderd}

\item{M}{the length of the sliding window}

\item{Nrec:}{number of components to be output}
}
\value{
a SSA object, compatible with the supplied plot function
}
\description{
Singular spectrum analysis as referred and described in Ghil et al.
and in the  \url{http://research.atmos.ucla.edu/tcd/ssa/}
results have been validated and shown to produce exactly the same results
as the Spectra toolking on 25.9.08
}
\examples{

bbridge <- function (n=256, amp) {
  rw1 <- cumsum(rnorm(n,1))*amp / (n)
  rw2 <- cumsum(rnorm(n,1))*amp / (n)
  rw2 <- rw2 * rw1[n] / rw2[n]
  bbridge <- rw1 - rw2
}

N <- 256

Phase1 <- bbridge(N,2*pi)
Phase2 <- bbridge(N,2*pi)
Amp1 <- 1 + bbridge(N,0.5)
Amp2 <- 1 + bbridge(N,0.5)

t = seq(N)

P1 = 40
P2 = 90

sig1 <- sin(2*pi*t/P1 + Phase1)*Amp1
sig2 <- sin(2*pi*t/P2 + Phase2)*Amp2

noise <- bbridge(N, 0.3) + rnorm(N,0.2)

signal <- noise + sig1 + sig2

plot (signal, type='l')

M <- 50

SSAsignal <- ssa (signal, M=M, Nrec=M)

NRec <- c(10,30,M)

for (i in seq(along=NRec)) {
rec <- apply(SSAsignal$PCA[seq(NRec[i]),], 2, sum)
lines(rec,col=i+1)
}

# full reconstruction (modulo the mean of the signal)
fullrec <- apply(SSAsignal$PCA[seq(M),], 2, sum)

# show that the difference is small
plot(fullrec - signal + mean(signal))

}
\references{
\insertRef{Ghil02aa}{gtseries}
\insertRef{VAUTARD89aa}{gtseries}
}
